update folder name from benefits_documents to evidence_submissions
