+++ modules/mobile/spec/requests/mobile/v0/claim/documents_spec.rb
@@ -34,12 +34,12 @@ RSpec.describe 'Mobile::V0::Claim::Document', :skip_json_api_validation, type: :
     params = { file:, claim_id:, tracked_item_id:, document_type: }
     expect do
       post '/mobile/v0/claim/600117255/documents', params:, headers: sis_headers
-    end.to change(Lighthouse::BenefitsDocuments::DocumentUpload.jobs, :size).by(1)
+    end.to change(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs, :size).by(1)
 
     expect(response).to have_http_status(:accepted)
     expect(response.parsed_body.dig('data',
-                                    'jobId')).to eq(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['jid'])
-    expect(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first.dig('args', 1,
+                                    'jobId')).to eq(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['jid'])
+    expect(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first.dig('args', 1,
                                                                         'tracked_item_id')).to eq([tracked_item_id])
   end
 
@@ -51,11 +51,11 @@ RSpec.describe 'Mobile::V0::Claim::Document', :skip_json_api_validation, type: :
     expect do
       post '/mobile/v0/claim/600117255/documents/multi-image', params: params.to_json,
                                                                headers: sis_headers(json: true)
-    end.to change(Lighthouse::BenefitsDocuments::DocumentUpload.jobs, :size).by(1)
+    end.to change(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs, :size).by(1)
     expect(response).to have_http_status(:accepted)
     expect(response.parsed_body.dig('data',
-                                    'jobId')).to eq(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['jid'])
-    expect(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first.dig('args', 1,
+                                    'jobId')).to eq(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['jid'])
+    expect(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first.dig('args', 1,
                                                                         'tracked_item_id')).to eq([tracked_item_id])
   end
 
@@ -64,11 +64,11 @@ RSpec.describe 'Mobile::V0::Claim::Document', :skip_json_api_validation, type: :
       params = { file:, claimId: claim_id, trackedItemId: tracked_item_id, documentType: document_type }
       expect do
         post '/mobile/v0/claim/600117255/documents', params:, headers: sis_headers(camelize: false)
-      end.to change(Lighthouse::BenefitsDocuments::DocumentUpload.jobs, :size).by(1)
+      end.to change(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs, :size).by(1)
       expect(response).to have_http_status(:accepted)
       expect(response.parsed_body.dig('data',
-                                      'job_id')).to eq(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['jid'])
-      expect(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first.dig('args', 1,
+                                      'job_id')).to eq(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['jid'])
+      expect(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first.dig('args', 1,
                                                                           'tracked_item_id')).to eq([tracked_item_id])
     end
 
@@ -80,12 +80,12 @@ RSpec.describe 'Mobile::V0::Claim::Document', :skip_json_api_validation, type: :
       expect do
         post '/mobile/v0/claim/600117255/documents/multi-image', params: params.to_json,
                                                                  headers: sis_headers(camelize: false, json: true)
-      end.to change(Lighthouse::BenefitsDocuments::DocumentUpload.jobs, :size).by(1)
+      end.to change(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs, :size).by(1)
 
       expect(response).to have_http_status(:accepted)
       expect(response.parsed_body.dig('data',
-                                      'job_id')).to eq(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['jid'])
-      expect(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first.dig('args', 1,
+                                      'job_id')).to eq(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['jid'])
+      expect(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first.dig('args', 1,
                                                                           'tracked_item_id')).to eq([tracked_item_id])
     end
   end
@@ -99,10 +99,10 @@ RSpec.describe 'Mobile::V0::Claim::Document', :skip_json_api_validation, type: :
     expect do
       post '/mobile/v0/claim/600117255/documents/multi-image', params: params.to_json,
                                                                headers: sis_headers(json: true)
-    end.to change(Lighthouse::BenefitsDocuments::DocumentUpload.jobs, :size).by(1)
+    end.to change(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs, :size).by(1)
     expect(response).to have_http_status(:accepted)
     expect(response.parsed_body.dig('data',
-                                    'jobId')).to eq(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['jid'])
+                                    'jobId')).to eq(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['jid'])
   end
 
   it 'uploads multiple mixed img files' do
@@ -113,10 +113,10 @@ RSpec.describe 'Mobile::V0::Claim::Document', :skip_json_api_validation, type: :
     expect do
       post '/mobile/v0/claim/600117255/documents/multi-image', params: params.to_json,
                                                                headers: sis_headers(json: true)
-    end.to change(Lighthouse::BenefitsDocuments::DocumentUpload.jobs, :size).by(1)
+    end.to change(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs, :size).by(1)
     expect(response).to have_http_status(:accepted)
     expect(response.parsed_body.dig('data',
-                                    'jobId')).to eq(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['jid'])
+                                    'jobId')).to eq(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['jid'])
   end
 
   it 'rejects files with invalid document_types' do
@@ -131,10 +131,10 @@ RSpec.describe 'Mobile::V0::Claim::Document', :skip_json_api_validation, type: :
   it 'normalizes requests with a null tracked_item_id' do
     params = { file:, claim_id:, tracked_item_id: nil, document_type: }
     post '/mobile/v0/claim/600117255/documents', params:, headers: sis_headers
-    args = Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['args'][1]
+    args = Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['args'][1]
     expect(response).to have_http_status(:accepted)
     expect(response.parsed_body.dig('data',
-                                    'jobId')).to eq(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['jid'])
+                                    'jobId')).to eq(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['jid'])
     expect(args.key?('tracked_item_id')).to eq(true)
     expect(args['tracked_item_id']).to be_nil
   end
@@ -148,11 +148,11 @@ RSpec.describe 'Mobile::V0::Claim::Document', :skip_json_api_validation, type: :
       params = { file:, claim_id:, tracked_item_id:, document_type: }
       expect do
         post '/mobile/v0/claim/600117255/documents', params:, headers: sis_headers
-      end.to change(Lighthouse::BenefitsDocuments::DocumentUpload.jobs, :size).by(1)
+      end.to change(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs, :size).by(1)
 
       expect(response).to have_http_status(:accepted)
       expect(response.parsed_body.dig('data',
-                                      'jobId')).to eq(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['jid'])
+                                      'jobId')).to eq(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['jid'])
     end
   end
 
@@ -183,7 +183,7 @@ RSpec.describe 'Mobile::V0::Claim::Document', :skip_json_api_validation, type: :
       post '/mobile/v0/claim/600117255/documents', params:, headers: sis_headers
       expect(response).to have_http_status(:accepted)
       expect(response.parsed_body.dig('data',
-                                      'jobId')).to eq(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['jid'])
+                                      'jobId')).to eq(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['jid'])
     end
 
     it 'rejects locked PDFs with the incorrect password' do
@@ -257,7 +257,7 @@ RSpec.describe 'Mobile::V0::Claim::Document', :skip_json_api_validation, type: :
       post '/mobile/v0/claim/600117255/documents', params:, headers: sis_headers
       expect(response).to have_http_status(:accepted)
       expect(response.parsed_body.dig('data',
-                                      'jobId')).to eq(Lighthouse::BenefitsDocuments::DocumentUpload.jobs.first['jid'])
+                                      'jobId')).to eq(Lighthouse::EvidenceSubmissions::DocumentUpload.jobs.first['jid'])
     end
   end
